+++ CHANGELOG.md
@@ -11,17 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- `ModuleFile::has_signature` for compressed modules
+- `ModuleFile::has_signature` for compressed modules.
 - `ModuleFile::info` for modules without parameters. Previously it would panic.
+- `ModuleFile::info` for modules without parameter descriptions. Previously it would panic.
+- `ModParam::description` type changed to `Option<String>`.
 
 ### Added
 
-- Error handling
+- Error handling.
 
 ### Changed
 
 - `ModuleFile` methods now return errors instead of panicking.
-- `ModuleFile::info` returns `&ModInfo` instead of `ModInfo`
+- `ModuleFile::info` returns `&ModInfo` instead of `ModInfo`.
 
 ## [0.3.4] - 2020-02-17