+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ModuleFile::has_signature` for compressed modules.
 - `ModuleFile::info` for modules without parameters. Previously it would panic.
 - `ModuleFile::info` for modules without parameter descriptions. Previously it would panic.
-- `ModParam::description` type changed to `Option<String>`.
+- Actually implement `LoadedModule::parameters`. Oops.
 
 ### Added
 
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ModuleFile` methods now return errors instead of panicking.
 - `ModuleFile::info` returns `&ModInfo` instead of `ModInfo`.
+- `ModParam::description` type changed to `Option<String>`.
 
 ## [0.3.4] - 2020-02-17