+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.4.0] - 2020-02-18
+
 ### Fixed
 
 - `ModuleFile::has_signature` for compressed modules.
@@ -123,7 +125,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release. No notable features
 
 <!-- next-url -->
-[Unreleased]: https://github.com/DianaNites/linapi/compare/v0.3.4...HEAD
+[Unreleased]: https://github.com/DianaNites/linapi/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/DianaNites/linapi/compare/v0.3.4...v0.4.0
 [0.3.4]: https://github.com/DianaNites/linapi/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/DianaNites/linapi/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/DianaNites/linapi/compare/v0.3.1...v0.3.2